Update to_sql command to use append instead of replace
The original version was using 'replace' to instruct pandas to drop
and recreate the postgres table before loading the data. This made
the script re-runnable but didn't do a great job of proving out
the data cleanup. Instead, use 'append' to respect the original
schema constraints (and keep things re-runnable by deleting the
rows before the bulk insert).